+++ changelog.md
@@ -1,6 +1,7 @@
 ## Versions
 * [0.1.0](https://github.com/arushadev/piraye/releases/tag/0.1.0)
   * change tokenizer to nltk
+  * pass custom tokenizer to normalizer
 * [0.0.3](https://github.com/arushadev/piraye/releases/tag/0.0.3)
     * fixed some bugs
 * [0.0.2](https://github.com/arushadev/piraye/releases/tag/0.0.2)